+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v0.0.0 2026-03-06
+## v0.0.1 2026-03-06
 
 Added:
 
+- added installers with support for ollama
+- added devcontainer support
 - added RAG
 Removed:
 
 Improved/Fixed:
+- fixed installer for windows hanging on ollama